backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
python-dotenv
db-dtypes
google-cloud-geminidataanalytics
